+++ src/components/Hero/Hero.stories.tsx
@@ -54,7 +54,7 @@ const otherSubHeaderTextLong = (
 );
 const imageProps = {
   alt: "Image example",
-  src: "//placekitten.com/800/400",
+  src: "https://iiif.nypl.org/iiif/2/5164274/full/!900,900/0/default.jpg",
 };
 
 const meta: Meta<typeof Hero> = {
@@ -107,7 +107,7 @@ export const WithControls: Story = {
     (args.heroType === "primary" && (
       <Hero
         {...args}
-        backgroundImageSrc="//placekitten.com/2400/800"
+        backgroundImageSrc="https://iiif.nypl.org/iiif/2/5164274/full/!900,900/0/default.jpg"
         heading={
           <Heading
             level="h1"
@@ -155,7 +155,7 @@ export const WithControls: Story = {
     (args.heroType === "campaign" && (
       <Hero
         {...args}
-        backgroundImageSrc="//placekitten.com/2400/800"
+        backgroundImageSrc="https://iiif.nypl.org/iiif/2/5164274/full/!900,900/0/default.jpg"
         heading={<Heading level="h1" id="1" text="Hero Campaign" />}
         heroType={args.heroType}
         imageProps={args.imageProps}
@@ -168,7 +168,7 @@ export const WithControls: Story = {
         heroType={args.heroType}
         imageProps={{
           ...args.imageProps,
-          src: "//placekitten.com/1200/400",
+          src: "https://iiif.nypl.org/iiif/2/5164274/full/!900,900/0/default.jpg",
         }}
         subHeaderText={
           <>
@@ -190,7 +190,7 @@ export const WithControls: Story = {
 export const Primary: Story = {
   render: () => (
     <Hero
-      backgroundImageSrc="//placekitten.com/1600/800"
+      backgroundImageSrc="https://iiif.nypl.org/iiif/2/5164274/full/!900,900/0/default.jpg"
       heading={
         <Heading
           id="primary-hero"
@@ -261,7 +261,7 @@ export const Campaign: Story = {
           text="Campaign Hero at Default Height"
         />
         <Hero
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc="https://iiif.nypl.org/iiif/2/5164274/full/!900,900/0/default.jpg"
           heroType="campaign"
           heading={
             <Heading
@@ -270,7 +270,10 @@ export const Campaign: Story = {
               text="Hero Campaign"
             />
           }
-          imageProps={imageProps}
+          imageProps={{
+            alt: "",
+            src: "https://iiif.nypl.org/iiif/2/5164274/full/!900,900/0/default.jpg",
+          }}
           subHeaderText={otherSubHeaderText}
         />
       </div>
@@ -281,7 +284,7 @@ export const Campaign: Story = {
           text="Campaign Hero with Long Text"
         />
         <Hero
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc="https://iiif.nypl.org/iiif/2/5164274/full/!900,900/0/default.jpg"
           heroType="campaign"
           heading={
             <Heading
@@ -304,7 +307,7 @@ export const CampaignDarkBackgroundImage: Story = {
     <Stack spacing="l">
       <div>
         <Hero
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc="https://iiif.nypl.org/iiif/2/5164274/full/!900,900/0/default.jpg"
           heroType="campaign"
           heading={
             <Heading